# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quick Turn",
        "description": "Documentación de API QuickTurn",
        "contact": {
            "email": "germanca8@correo.ugr.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/entidads": {
            "get": {
                "tags": [
                    "Entidades"
                ],
                "summary": "Lista todas las entidades",
                "description": "Endpoint para listar las entidades",
                "operationId": "af183242f5116da6bbbf4bb9e173cb8a",
                "responses": {
                    "200": {
                        "description": "Lista todas las entidades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Entidad"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Entidades"
                ],
                "summary": "Crea una entidad",
                "description": "Endpoint para almacenar una entidad",
                "operationId": "d43f9fa19a6c72fefbb3a2da23d878d2",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreEntidad"
                },
                "responses": {
                    "200": {
                        "description": "Entidad creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entidad"
                                }
                            }
                        }
                    },
                    "444": {
                        "$ref": "#/components/responses/444"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/entidads/{entidadID}": {
            "get": {
                "tags": [
                    "Entidades"
                ],
                "summary": "Muestra una entidad por ID",
                "description": "Endpoint para mostrar una entidad",
                "operationId": "50f97ed6b2cf7ce4bcab86c008919785",
                "parameters": [
                    {
                        "name": "entidadID",
                        "in": "path",
                        "description": "ID de la entidad a obtener",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la entidad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entidad"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Entidades"
                ],
                "summary": "Elimina una entidad",
                "description": "Endpoint para eliminar una entidad",
                "operationId": "df04696ecc5eb575caf2b9fe498c9ea9",
                "parameters": [
                    {
                        "name": "entidadID",
                        "in": "path",
                        "description": "ID de la entidad a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entidad eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Entidades"
                ],
                "summary": "Actualiza una entidad",
                "description": "Endpoint para actualizar una entidad",
                "operationId": "87b193862ebc772296b93f68d7e0e592",
                "parameters": [
                    {
                        "name": "entidadID",
                        "in": "path",
                        "description": "ID de la entidad a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateEntidad"
                },
                "responses": {
                    "200": {
                        "description": "Entidad actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entidad"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/entidads/users": {
            "post": {
                "tags": [
                    "Entidades"
                ],
                "summary": "Lista todas las entidades de 1 usuario",
                "description": "Endpoint para mostrar todas las entidades asociadas a un usuario",
                "operationId": "b4ced2a29f48d7b70615567255bd5e96",
                "requestBody": {
                    "$ref": "#/components/requestBodies/EntidadRequest"
                },
                "responses": {
                    "200": {
                        "description": "Lista todas las entidades de 1 usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Entidad"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Entidad": {
                "properties": {
                    "id": {
                        "description": "Id asociado",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "user_id": {
                        "description": "Id del usuario",
                        "type": "string",
                        "example": "670b9562-b30d-52d5-b827-655787665500"
                    },
                    "nombreEntidad": {
                        "description": "Nombre de la entidad",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Mercadona"
                    },
                    "direccionEntidad": {
                        "description": "Direccion de la entidad",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Calle Colon"
                    },
                    "horarioEntidad": {
                        "description": "Horario de la entidad",
                        "type": "string",
                        "maxLength": 255,
                        "example": "9h - 21h"
                    },
                    "nombreFiscal": {
                        "description": "Nombre legal de la entidad",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Mercadona SA"
                    },
                    "nif": {
                        "description": "Documento oficial de la entidad",
                        "type": "string",
                        "default": null,
                        "maxLength": 255,
                        "example": "11122233P"
                    },
                    "url": {
                        "description": "URL de la web de la entidad",
                        "type": "string",
                        "example": "https://www.mercadona.es/"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "example": "2022-03-15T13:23:02.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última modificación",
                        "type": "string",
                        "example": "2022-03-15T13:23:02.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "404": {
                "description": "No query results for model",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "444": {
                "description": "Name its already used",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Name is already used"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "requestBodies": {
            "EntidadRequest": {
                "description": "ID de la entidad",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "properties": {
                                "id": {
                                    "$ref": "#/components/schemas/Entidad/properties/id"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "StoreEntidad": {
                "description": "Entidad que necesita ser creada",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "user_id",
                                "nombreEntidad"
                            ],
                            "properties": {
                                "user_id": {
                                    "$ref": "#/components/schemas/Entidad/properties/user_id"
                                },
                                "nombreEntidad": {
                                    "$ref": "#/components/schemas/Entidad/properties/nombreEntidad"
                                },
                                "direccionEntidad": {
                                    "$ref": "#/components/schemas/Entidad/properties/direccionEntidad"
                                },
                                "horarioEntidad": {
                                    "$ref": "#/components/schemas/Entidad/properties/horarioEntidad"
                                },
                                "nombreFiscal": {
                                    "$ref": "#/components/schemas/Entidad/properties/nombreFiscal"
                                },
                                "nif": {
                                    "$ref": "#/components/schemas/Entidad/properties/nif"
                                },
                                "url": {
                                    "$ref": "#/components/schemas/Entidad/properties/url"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateEntidad": {
                "description": "Entidad que necesita ser actualizada",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "nombreEntidad"
                            ],
                            "properties": {
                                "nombreEntidad": {
                                    "$ref": "#/components/schemas/Entidad/properties/nombreEntidad"
                                },
                                "direccionEntidad": {
                                    "$ref": "#/components/schemas/Entidad/properties/direccionEntidad"
                                },
                                "horarioEntidad": {
                                    "$ref": "#/components/schemas/Entidad/properties/horarioEntidad"
                                },
                                "nombreFiscal": {
                                    "$ref": "#/components/schemas/Entidad/properties/nombreFiscal"
                                },
                                "nif": {
                                    "$ref": "#/components/schemas/Entidad/properties/nif"
                                },
                                "url": {
                                    "$ref": "#/components/schemas/Entidad/properties/url"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum Token Auth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}